Nested loop

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,7 +8,7 @@ int main(void)
 {
 	int t1 = 1;
 	int t2 = 2;
-	int next = t1 + t2;
+	long next = t1 + t2;
 	int num = 50 - 2;/* to remove first two values from the total loop*/
 
 	printf("%d, %d, ", t1, t2);
